pyproject.toml
[project]
name = "fractal"
version = "0.1.0"
description = "Self-similar agent decomposition - Build complex agent systems through elegant delegation"
readme = "README.md"
requires-python = ">=3.9"
keywords = ["ai", "agents", "delegation", "observability", "multi-agent", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
fastapi = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
